refactor: move operation types to SharpFM.Model
