+++ src/SharpFM.Model/Schema/FieldOperation.cs
@@ -1,9 +1,4 @@
-// This file is part of SharpFM and is licensed under the GNU General Public License v3.
-//
-// Plugin Exception: You may create plugins that implement these interfaces without those
-// plugins being subject to the GPL. Such plugins may use any license, including proprietary.
-
-namespace SharpFM.Plugin;
+namespace SharpFM.Model.Schema;
 
 /// <summary>
 /// A single field mutation within a batch update to a table clip.
