+++ src/SharpFM.Model/Scripting/ScriptStepOperation.cs
@@ -1,11 +1,6 @@
-// This file is part of SharpFM and is licensed under the GNU General Public License v3.
-//
-// Plugin Exception: You may create plugins that implement these interfaces without those
-// plugins being subject to the GPL. Such plugins may use any license, including proprietary.
-
 using System.Collections.Generic;
 
-namespace SharpFM.Plugin;
+namespace SharpFM.Model.Scripting;
 
 /// <summary>
 /// A single script step mutation within a batch update.
